+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixes
 
+- Fixed error JSON.parse while passing both `returnAs:'BUFFER'` and `allowedStatusCode:[204]`
 - Fixed PostDecoratorDecoratedRequest typing
 
 ## v4.2.0 - 2021-10-15